Move save_regions function to new tasks/regions.py
